Annoyingly handhold the snippet ordering.
Once the typemapper is able to recognize uses of a UDT that is
being defined, it should also certainly be able to enter a
provides-requires relationship between the type and any SQL snippet
that uses it. Until then, these need to be added explicitly, to
make sure things are written to the DDR in a workable order.