Annotations can supply arbitrary SQL too.
Declarations of the min and max aggregates have been moved into
RationalUDT.java to be near their code. Provides/requires tags
are used to make sure they will not be emitted before the
functions they depend on.

The same process could be followed to eventually move all of the
SQL declarations into the Java sources, and have no more need for
the hand-maintained DDR in addition to the generated one. Or the
partly-generated, partly hand-maintained approach could be kept
indefinitely, to avoid the tedium of moving all of those operators
over (with provides/requires tags, of course, to get the order right).

Perhaps it will seem more attractive to wait and see if some future
PL/Java release adds annotations to automatically generate the
operator/class/family declarations with the right code in the right
order, and finish the transition then.

This SQLAction block also supplies remove actions to drop the
aggregates. That could be seen as unnecessarily obsessive (the
DROP TYPE ... CASCADE at the end will handle it, after all), but
also illustrates how to do that.