There can be multiple SQLAction blocks.
In all supported Java versions, they can be put inside
an SQLActions (note the s) block. Starting with Java 8,
PL/Java could (but does not yet) support the newly added
ability to repeat the same annotation without wrapping it
in a containing annotation. It doesn't seem terribly
urgent, as (a) it is only doing exactly the same thing
under the hood, and (b) it would unnecessarily encourage
writing code that couldn't work in Java 6 or 7.