Cull duplicated functions.
What do these three methods have in common? (a) they are the same as
other methods that take primitive parameters, except that these take
the boxed form instead; (b) the usefulness of declaring the boxed form
of a parameter is that you can accept a null value for the parameter,
and recognize it in the familiar Java way; and (c) all three of these
methods have been annotated RETURNS_NULL onNullInput, meaning they will
never be called with null parameter values, because PostgreSQL's query
optimizer would just use NULL as the result without calling them at all.

So, these are superfluous, and the best way to resolve the duplicate-
function-declaration issues they would cause is to not have them.