Annoyingly handhold the typemapper.
Here is a current limitation of the SQL generator: it is generating
this UDT, it has all the information it would need to recognize that
type when used in function or parameter declarations, but it does not,
and reports "no known mapping to an SQL type". Until that is fixed
in a later release, it has to be helped along by adding more annotations:
@SQLType on the parameters, or type= in the @Function annotation when
the hint is needed for the function's return type.