There can be two deployment descriptors.
The autogenerated one so far takes care of the types and
associated functions. Everything declared there can be removed
from the hand-curated one--which already reveals why generating
is safer than hand-curating: some function declarations in the
hand-built one turn out not to correspond to anything in the
Java source! That also illustrates the benefit of having PL/Java
implement a 'validator' entry point; without it, those function
declarations can seem to succeed, with no problem apparent
until the first attempts to call them.

When there are multiple deployment descriptors, PL/Java processes
them in the order with which they are declared in the jar manifest
(that's what the SQL/JRT standard specifies). The generated one,
pljava.ddr, defines the types and functions needed by what is left
in the hand-built one, so pljava.ddr is mentioned first in the POM.

Many Java tools do not guarantee to preserve order of manifest
sections (Java's jar specification says order doesn't matter,
the SQL/JRT standard says for this purpose it does). Will the
maven-jar-plugin always write these entries into the manifest in
their order of appearance in the POM? I am not sure; a paranoid
build procedure might wisely include a test. But in several tries
in a row right here and now, it did the right thing.

Things can be commented in a hand-crafted DDR. Owing to
limitations in the code that reads it, the /* ... */ comment
style is best.

There is still one thing wrong with this pair of DDRs.