Use PL/Java 1.6.0 to save more tedium.
PL/Java 1.6.0's SQL generator is able to order UDTs and functions
it generates so that manually-added provides/requires are not needed
to make sure a UDT is created before functions that refer to it as
parameter or return type. By bumping the pljava-api dependency to
1.6.0, many of the annotations tediously added in commit d32f84e
for earlier PL/Java versions can now be avoided.

With PL/Java 1.6.0, the minimum supported Java release is 9, so edit
the POM to use a newer maven-compiler-plugin and specify 9 as the
Java release.