Force resource bundles to FORMAT_PROPERTIES
There is a curious reason why these examples fail to load
into PL/Java with exception-in-initializer errors.

Each one tries to load a ResourceBundle, and uses Class.getName()
as the base name of the bundle. Indeed, the jar includes files
with pathnames that exactly match the class files, but for the
extension .properties in place of .class.

The thing is, the getBundle(basename) that takes only a basename
tries to find it in the formats in ResourceBundle.Control.FORMAT_DEFAULT,
which lists "java.class" and "java.properties" in that order. So the
first thing getBundle tries to do is look for a class with the exact
name just obtained from Class.getName()! Sure enough, it finds one!

Of course the class it finds isn't a subclass of ResourceBundle,
so that excitement dies down quickly with a ClassCastException.

I am not sure how this was even originally meant to work. Checking
Java's API docs, the order of FORMAT_DEFAULT has been class first,
then properties, at least as far back as Java 6.

Solved here in the laziest way, just by adding a Control argument
to force FORMAT_PROPERTIES. However, that will lead to a future
headache somewhere down the road: when a future PL/Java supports
modular user code and these examples can be made into modules ...
ResourceBundle.Control is not allowed in modular code. But that
bridge can be crossed when it comes.

Another solution would be to give the property files names that
simply don't exactly match the class names.

In passing, also fixed what appeared to be a copy-pasto: in each
of the two classes, the class object getName() was called on was
the one for ComplexUDT, even though there is a property file
for each.