Use synthetic operators in RationalUDT
The number of boilerplate methods having to be written to support
btree comparison strategies can be reduced starkly by just writing,
say, a method for < and one for = and declaring their commutators
and negators in Operator annotations with the synthetic= attribute.

PL/Java will emit actual function declarations at the SQL level
(they will refer to the base Java methods, with commutation and/or
negation applied), so the synthetic= attribute takes the name that
should be given to the synthetic SQL function.

The technique isn't taken as far in this commit as it could be;
the actual Java methods lessThanOrEquals and greaterThanOrEquals
are kept around, because they are used by other methods in
the Java class.