Use synthetic operators in ComplexUDT
Synthetic operators can also save writing redundant Java methods
to implement the commutators of cross-type operators like + or *.

Unlike the comparison operators in the previous commit (where
commuting a < operator changes the meaning and makes it a > operator),
operators like + and * are truly commutative, so it makes perfect
sense to give the synthesized function the very same name (complex_add
or complex_multiply here) as the one it is based on. PostgreSQL
allows that, provided the operand types are different, as that means
the commuted version will have a different signature.

The shorthand notation SELF, used earlier, isn't quite right for that;
a new one, TWIN, will be introduced in a later commit to make
this case tidier.