Eliminate an = method with 1.6.3's synthesis
Bump the PL/Java API dependency to 1.6.3 to eliminate one of those
hand-coded cross-type = implementations for RationalUDT and double.
As long as = and <> are both commutative, the operators and their
commutators can be declared on one implementing method, using
commutator = TWIN and synthetic = TWIN.

This way, it's only necessary to implement three methods to get the
full complement of cross-type comparisons: one of < <= >= > for each
of A,B and B,A, and just one of = <> for either A,B or B,A. Nine
other operators can be synthesized from those.